change game played via prompt
Choice is received via prompt. Also change UI to a much more
simple layout. NOTE: their does seem to be a bug where the
website does not load or something and it seems to not
be able to "remember" the score. Ex. 5 rounds are played
and only 4 points are awarded.